refactor: split main code emitter from its dependencies
